Some refacto and add check_user input
