+++ Dockerfile
@@ -3,6 +3,7 @@ FROM python:3
 WORKDIR /usr/src/app
 
 COPY requirements ./
+COPY tox.ini ./
 RUN pip install --no-cache-dir -r requirements
 
 COPY ./src/ ./
